requirements.txt
transformers==4.43.1
torch==2.5.0
coremltools==8.3.0
accelerate==0.27.2
python-dotenv>=1.0.0
sentencepiece==0.1.99
protobuf==3.20.3
